+++ .storybook/blocks/Example.tsx
@@ -7,6 +7,7 @@ export const Example = ({
   of,
   children,
 }: PropsWithChildren<{ of: ModuleExport }>) => {
+  console.log({ of });
   return (
     <>
       <StoryTitle of={of} />
