--- src/components/list/index.ts
@@ -1,1 +1,0 @@
-export * from "./List";
