+++ src/components/nested.tsx
@@ -3,7 +3,7 @@ import type { FormFields } from "form-atoms";
 import type { RenderProp } from "react-render-prop-type";
 
 import type { ListAtom } from "../atoms";
-import { type Components, createComponents } from ".";
+import { type Components, createList } from "./list";
 
 export type NestedProps<Fields extends FormFields> = {
   atom: ListAtom<Fields, any>;
@@ -13,7 +13,7 @@ export function Nested<Fields extends FormFields>({
   atom,
   children,
 }: NestedProps<Fields>) {
-  const [components] = useState(() => createComponents(atom));
+  const [components] = useState(() => createList(atom));
 
   return children(components as any);
 }
